Silenced a visual studio warning

+++ src/CheckSum.cpp
@@ -237,7 +237,7 @@ inline static const uint8_t *unfoldedMapSegs()
     
     for(size_t i = 0; i < MAPPLANES; ++i)
     {
-        memcpy(mapSegsLinear + i * mapSegsUnitLength, mapSegs[i],
+        memcpy(mapSegsLinear + i * mapSegsUnitLength, mapSegs[(int)i],
                mapSegsUnitLength);
     }
